fixed warns being global, working on subcommands

+++ commands/purge.js
@@ -1,4 +1,4 @@
-const { SlashCommandBuilder } = require('discord.js');
+const { SlashCommandBuilder, PermissionsBitField, PermissionFlagsBits } = require('discord.js');
 
 module.exports = {
 	data: new SlashCommandBuilder()
@@ -8,17 +8,23 @@ module.exports = {
 			option
 				.setName('messages')
 				.setDescription('how many messages to delete.')
-				.setRequired(true)),
+				.setRequired(true))
+		.setDefaultMemberPermissions(PermissionFlagsBits.ManageMessages),
 	async execute(interaction) {
 		//define first response
-        const numMessages = interaction.options.getInteger("messages")
-		//then update message with ping
-        interaction.channel.bulkDelete(numMessages).then(() => {
-            interaction.reply({ content: `Deleted ${numMessages} messages.`, ephemeral: true })
-          }).catch(error => {
-            console.error(error);
-            interaction.reply('Error deleting messages.');
-          });
+		if (interaction.member.permissions.has(PermissionsBitField.Flags.ManageMessages)) {
+			const numMessages = interaction.options.getInteger("messages")
+			//then update message with ping
+			interaction.channel.bulkDelete(numMessages).then(() => {
+				interaction.reply({ content: `Deleted ${numMessages} messages.`, ephemeral: true })
+			  }).catch(error => {
+				console.error(error);
+				interaction.reply('Error deleting messages.');
+			  });
+		} else {
+			interaction.reply(`You don't have permission to run this!`);
+		}
+
 		console.log('Purge command - completed')
 	},
 };
